UI/requirements.txt
absl-py==1.4.0
altair==5.0.1
anyio==3.7.0
astor==0.8.1
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.83
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.26.165
botocore==1.29.165
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cssselect==1.2.0
cssutils==2.7.1
cycler==0.10.0
Cython==0.29.35
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
flask-babel==3.1.0
fonttools==4.40.0
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==2.10
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
lazy_loader==0.3
lmdb==1.4.1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.5.0
pandas==2.0.3
pdf2docx==0.5.6
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
premailer==3.10.0
protobuf==3.20.2
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pyclipper==1.3.0.post4
pycryptodome==3.18.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
PyMuPDF==1.20.2
pyparsing==2.4.7
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==3.1.1
rarfile==4.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.4.2
roboflow==1.1.0
rsa==4.9
s3transfer==0.6.1
sahi==0.11.14
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
streamlit==1.24.0
supervision==0.11.1
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tifffile==2023.4.12
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.7.0
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.124
urllib3==1.26.16
validators==0.20.0
visualdl==2.5.3
watchdog==3.0.0
Werkzeug==2.3.6
wget==3.2
yolov5==7.0.12
zipp==3.15.0
